refactor(codeScan): extract sleepWithAbort to shared util
Move sleepWithAbort from scanner/request.ts to util/time.ts
to avoid code duplication and make it available for other use cases.
